% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAAChangeTable.R
\name{makeAAChangeTable}
\alias{makeAAChangeTable}
\title{Make Amino Acid Change Table}
\usage{
makeAAChangeTable(maf, riskcat, gene, csv = FALSE)
}
\arguments{
\item{maf}{Data frame returned by GDCquery_Maf}

\item{riskcat}{Data frame containing the bspcacat risk classifier for each patient}

\item{gene}{Gene name (Hugo symbol) as a string}

\item{csv}{Logical that tells the function whether to write the result to a csv file or not. Defaults to FALSE}
}
\value{
Data frame containing computed frequencies of each amino acid change in high, intermediate, and low risk patients.
}
\description{
This function takes in the MAF data frame produced by GDCquery_Maf from TCGAbiolinks, a clinical information
table with the bspcacat risk classifer, and a gene name (Hugo symbol).
It return a data frame of the frequencies of each amino acid change in
high, intermediate, and low risk of recurrence patients, in addition to all Stage I LUAD patients and all LUAD
patients from the TCGA database.
Optionally, if csv is set to TRUE, it writes the data frame to a csv file in the folder 'AAChanges' in the current
directory.
}
