% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAChangeTable.R
\name{plotAAChangeTable}
\alias{plotAAChangeTable}
\title{Plot Amino Acid Change Table}
\usage{
plotAAChangeTable(table, minfreq = 2)
}
\arguments{
\item{table}{Data frame produced as output from makeAAChangeTable}

\item{minfreq}{Minimum frequency of amino acid changes from all Stage I patients to plot, defaults to 2.}
}
\description{
This function takes in the output data frame of makeAAChangeTable and plots it in bar graph form.
By default, only plots amino acid changes with frequency in Stage I patients greater than 1.
This can be changed by setting the minfreq parameter.
}
